Created folder utilities + moved utils.ts + migrated localStorageStore
