+++ frontend/src/utilities/utils.ts
@@ -2,10 +2,12 @@ type Method = 'GET' | 'POST' | 'PUT' | 'PATCH' | 'DELETE';
 
 /**
  * Get data from API
+ *
  * @param url url to fetch
  * @param data data which will be sent to the API
  * @param method HTTP method, if you want to override the default GET/POST
  * @note If data is passed, the request will be a POST request, otherwise a GET request, if not overridden by method parameter
+ *
  * @returns $ReturnType, if the request was successful, otherwise undefined
  */
 export const getFromAPI = async <$ReturnType>(
@@ -33,8 +35,10 @@ export const getFromAPI = async <$ReturnType>(
 
 /**
  * Generate array with range of numbers from start
+ *
  * @param size size of the array
  * @param start starting number
+ *
  * @returns array of numbers from start to start + size
  */
 export const generateRange = (size: number, start = 0) => {
